Move `Tube` menu setting to `MapObjectPicker`
