+++ appOPHD/States/StructureTracker.h
@@ -5,16 +5,10 @@
 #include <vector>
 
 
-struct IconGridItem;
-
-
 class StructureTracker
 {
 public:
 	StructureTracker();
-
-	const std::vector<IconGridItem>& surfaceTubes() const;
-	const std::vector<IconGridItem>& undergroundTubes() const;
 
 	const std::vector<StructureID>& availableSurfaceStructures() const;
 	const std::vector<StructureID>& availableUndergroundStructures() const;
